requirements.txt
asgiref==3.4.1
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coverage==5.5
cryptography==3.3.1
defusedxml==0.7.0rc2
dj-database-url==0.5.0
Django==3.2.6
django-appconf==1.0.4
django-babel==0.6.2
django-blacklist==0.5.1
django-compressor==2.4
django-compressor-toolkit==0.6.0
django-crispy-forms==1.11.0
django-easy-select2==1.5.7
django-extensions==3.1.0
django-mathfilters==1.0.0
django-npm==1.0.0
django-ratelimit==3.0.1
django-webpack-loader==0.7.0
djangorestframework==3.12.2
ecdsa==0.17.0
freezegun==1.1.0
gunicorn==20.0.4
humanize==3.2.0
idna==2.10
lxml==4.6.2
mysql==0.0.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
Pillow==8.1.0
protobuf==3.15.5
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-jose==3.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytrends==4.7.3
pytz==2021.1
rcssmin==1.0.6
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.7.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
